Cuda renderer, sfloat image format

+++ abstract/IRenderer.hpp
@@ -5,5 +5,5 @@ class IRenderer
 {
 public:
 	virtual ~IRenderer() = default;
-	virtual void render(uint32_t* image_data, uint32_t width, uint32_t height) = 0;
+	virtual void render(float* image_data, uint32_t width, uint32_t height) = 0;
 };
